Move HAL related code to appropriate libraries
