+++ libraries/HAL/HAL.h
@@ -1,0 +1,8 @@
+//
+// Created by Anton Matosov on 5/6/16.
+//
+
+#pragma once 
+
+#import <SerialProtocol.h>
+#import <ProgramMemoryProtocol.h>
