+++ libraries/HALArduino/ArduinoHardwareSerial.cpp
